requirements.txt
certifi
feedparser
slackclient
